requirements.txt
# requirements.txt
mysql-connector-python
numpy
scipy
python-dotenv
Flask
requests
beautifulsoup4
selenium
